added methods to execute playback of sessions

See merge request VinteR/TheApplication!79

+++ VinteR/Model/Session.cs
@@ -22,7 +22,7 @@ namespace VinteR.Model
         /// <summary>
         /// Time in millis how long the session lasts
         /// </summary>
-        public long Duration { get; set; }
+        public uint Duration { get; set; }
 
         /// <summary>
         /// Frames for the session. This MUST NOT be set during a
